+++ tasks/sapo/sapo-web.config.ts
@@ -1,0 +1,2 @@
+export const SAPO_WEB_KEY = process.env.SAPO_WEB_KEY
+export const SAPO_WEB_SECRET = process.env.SAPO_WEB_SECRET
